added permissions, and necessary methods to get user location

+++ app/src/main/java/com/example/myforecast/ApiService.java
@@ -8,8 +8,8 @@ public interface ApiService {
 //    TODO - ask user which units he wants (metric,imperial)
     @GET("/data/2.5/forecast")
     Call<ForecastModel> getRecentForecast(
-            @Query("lat") int lat,
-            @Query("lon") int lon,
+            @Query("lat") double lat,
+            @Query("lon") double lon,
             @Query("appid") String api_key,
             @Query("units") String units);
 }
